+++ src/cpp/tests/test_tardigrade_hydraLinearElasticity.cpp
@@ -1855,7 +1855,7 @@ BOOST_AUTO_TEST_CASE( test_residual_setJacobian ){
 
     }
 
-    BOOST_CHECK( tardigradeVectorTools::fuzzyEquals( gradient, *R.getJacobian( ) ) );
+    BOOST_CHECK( tardigradeVectorTools::fuzzyEquals( tardigradeVectorTools::appendVectors( gradient ), *R.getJacobian( ) ) );
 
 }
 
@@ -2071,6 +2071,6 @@ BOOST_AUTO_TEST_CASE( test_residual_setdRdF ){
 
     }
 
-    BOOST_CHECK( tardigradeVectorTools::fuzzyEquals( gradient, *R.getdRdF( ) ) );
+    BOOST_CHECK( tardigradeVectorTools::fuzzyEquals( tardigradeVectorTools::appendVectors( gradient ), *R.getdRdF( ) ) );
 
 }