+++ src/cpp/tests/test_tardigrade_hydraLinearViscoelasticity.cpp
@@ -2257,7 +2257,7 @@ BOOST_AUTO_TEST_CASE( test_residual_setdRdT ){
 
     }
 
-    BOOST_CHECK( tardigradeVectorTools::fuzzyEquals( dRdF, *R.getdRdF( ) ) );
+    BOOST_CHECK( tardigradeVectorTools::fuzzyEquals( tardigradeVectorTools::appendVectors( dRdF ), *R.getdRdF( ) ) );
 
     BOOST_CHECK( tardigradeVectorTools::fuzzyEquals( dRdT, *R.getdRdT( ) ) );