+++ src/cpp/tests/test_tardigrade_hydraMicromorphicDruckerPragerPlasticity.cpp
@@ -17356,7 +17356,7 @@ BOOST_AUTO_TEST_CASE( test_setJacobian ){
 
     }
 
-    BOOST_CHECK( tardigradeVectorTools::fuzzyEquals( jacobian, *R.getJacobian( ) ) );
+    BOOST_CHECK( tardigradeVectorTools::fuzzyEquals( tardigradeVectorTools::appendVectors( jacobian ), *R.getJacobian( ) ) );
 
 }
 
@@ -17652,7 +17652,7 @@ BOOST_AUTO_TEST_CASE( test_setdRdD ){
 
     }
 
-    BOOST_CHECK( tardigradeVectorTools::fuzzyEquals( dRdD, *R.getdRdD( ) ) );
+    BOOST_CHECK( tardigradeVectorTools::fuzzyEquals( tardigradeVectorTools::appendVectors( dRdD ), *R.getdRdD( ) ) );
 
 }