+++ docs/sphinx/changelog.rst
@@ -38,6 +38,7 @@ Internal Changes
 - Removed all sayHello tests (:pull:`97`). By `Nathan Miller`_.
 - Improved performance of the linear elasticity subroutine (:pull:`99`). By `Nathan Miller`_.
 - Using new error_tools check for error function (:pull:`100`). By `Nathan Miller`_.
+- Changed Jacobian, dRdF, and dRdD to row-major vectors (:pull:`101`). By `Nathan Miller`_.
 
 Bug Fixes
 =========